refact: CourseRecommendService 클래스 SRP 적용
